+++ CHANGELOG.md
@@ -11,6 +11,12 @@ Changelog
 
 _Note: Gaps between patch versions are faulty, broken or test releases._
 
+## v3.73.6 (2022-01-31)
+
+#### :bug: Bug Fix
+
+* Fixed setting of the original prototype to the proxy object with the `accessors` engine `core/object/watch/engines/accessors`
+
 ## v3.74.5 (2022-01-25)
 
 #### :bug: Bug Fix